API-3276 undo package-lock.json changes, move tests, update paths
