+++ test/resolveSharedRefs/case-1/input.yaml
