+++ test/resolveSharedRefs/case-1/options.yaml
