+++ test/resolveSharedRefs/case-1/output.yaml
