+++ test/resolveSharedRefs/case-1/responses.yaml
