+++ test/resolveSharedRefs/case-1/schemas.yaml
