+++ test/resolveSharedRefs/case-1/schemas2.yaml
